+++ public/admin/_app/immutable/nodes/17.0OSqbwwy.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/9w5gwE7E.js";export{e as component};