+++ public/admin/_app/immutable/nodes/20.DW_eVvmA.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/DV0z3C3q.js";export{e as component};