+++ public/admin/_app/immutable/nodes/22.B7R-pNwY.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/D_uofSr9.js";export{e as component};