+++ public/admin/_app/immutable/nodes/28.ChsgXbAm.js
@@ -1,0 +1,1 @@
+import"../chunks/DOwKBK7S.js";import"../chunks/xihTtKlq.js";import"../chunks/C_L0dgGH.js";import{t as e}from"../chunks/9w5gwE7E.js";function t(t){e(t,{})}export{t as component};